requirements.txt
Flask==2.3.3
Pillow==10.0.0
opencv-python==4.8.0.76